Now All iam safe existing in the same the same file
- In aws-iam-safe file we can do the following:
    - Attach and detach iam safe to cloud account
    - Protect and unprotect iam safe entity
