+++ dome9_type_annotations/aws_cloud_account.py
@@ -1,7 +1,7 @@
 from typing import Dict, List
 
 from resources.aws_cloud_account import AwsCloudAccount as BaseAwsCloudAccount, AwsCloudAccountUpdateCredentials, AwsCloudAccountUpdateOrganizationalUnitID, \
- AwsCloudAccountUpdateConfig, AwsCloudAccountUpdateName, AwsCloudAccountRequest, AttachIamSafe, RestrictedIamEntitiesRequest
+ AwsCloudAccountUpdateConfig, AwsCloudAccountUpdateName, AwsCloudAccountRequest
 
 
 class aws_cloud_account(BaseAwsCloudAccount):
@@ -36,24 +36,4 @@ class aws_cloud_account(BaseAwsCloudAccount):
 
 	@classmethod
 	def delete(cls, aws_cloud_account_id: str):
-		pass
-
-	@classmethod
-	def attach_iam_safe(cls, body: AttachIamSafe) -> Dict:
-		pass
-
-	@classmethod
-	def detach_iam_safe(cls, aws_cloud_account_id: str) -> None:
-		pass
-
-	@classmethod
-	def protect_iam_safe_entity(cls, aws_cloud_account_id: str, body: RestrictedIamEntitiesRequest) -> str:
-		pass
-
-	@classmethod
-	def get_all_protected_iam_safe_entity(cls, aws_cloud_account_id: str) -> Dict:
-		pass
-
-	@classmethod
-	def unprotect_iam_safe_entity(cls, aws_cloud_account_id, entity_name: str, entity_type: str) -> None:
 		pass
