+++ dome9_type_annotations/client.py
@@ -3,6 +3,7 @@ from dome9_type_annotations.aws_cloud_account import aws_cloud_account
 from dome9_type_annotations.azure_cloud_account import azure_cloud_account
 from dome9_type_annotations.aws_security_group import aws_security_group
 from dome9_type_annotations.rule_bundle import rule_bundle
+from dome9_type_annotations.aws_iam_safe import aws_iam_safe
 
 
 class Client(BaseClient):
@@ -12,3 +13,4 @@ class Client(BaseClient):
 	aws_security_group: 'aws_security_group'
 	rule_bundle: 'rule_bundle'
 	continuous_compliance_notification: 'continuous_compliance_notification'
+	aws_iam_safe: 'aws_iam_safe'
